6 sns auth

+++ app/controllers/sign_ups_controller.rb
@@ -1,4 +1,5 @@
 class SignUpsController < ApplicationController
+  before_action :authenticate_user!, except: [:signup_select, :user_baseinfo, :user_baseinfo_validate, :user_tel, :user_tel_auth, :user_tel_validate, :user_params]
   before_action :user_params, only: :user_baseinfo_validate
   before_action :address_params, only: :user_address_create 
   before_action :credit_params, only: :user_credit_create
@@ -9,6 +10,8 @@ class SignUpsController < ApplicationController
 
   def user_baseinfo
     @user = User.new
+    session.delete(:provider)
+    session.delete(:uid)
     render layout: "application_sub"
   end
 
@@ -18,7 +21,7 @@ class SignUpsController < ApplicationController
     session[:birth]                 = user_params[:birth]
     session[:email]                 = user_params[:email]
     session[:password]              = user_params[:password]
-    session[:password_confirmation] = user_params[:password_confirmation]
+    session[:password_confirmation] = user_params[:password]
     session[:first_name]            = user_params[:first_name]
     session[:last_name]             = user_params[:last_name]
     session[:first_name_kana]       = user_params[:first_name_kana]
@@ -55,18 +58,17 @@ class SignUpsController < ApplicationController
 
   def user_tel_validate
     session[:tel] = params[:user][:tel]
-
     @user = User.new(
-      nickname:           session[:nickname],
-      birth:              session[:birth],
-      email:              session[:email],
-      password: session[:password],
+      nickname:              session[:nickname],
+      birth:                 session[:birth],
+      email:                 session[:email],
+      password:              session[:password],
       password_confirmation: session[:password_confirmation],
-      first_name:         session[:first_name],
-      last_name:          session[:last_name],
-      first_name_kana:    session[:first_name_kana],
-      last_name_kana:     session[:last_name_kana],
-      tel:                session[:tel],
+      first_name:            session[:first_name],
+      last_name:             session[:last_name],
+      first_name_kana:       session[:first_name_kana],
+      last_name_kana:        session[:last_name_kana],
+      tel:                   session[:tel],
     )
 
     @user.valid?
@@ -88,7 +90,7 @@ class SignUpsController < ApplicationController
       nickname:           session[:nickname],
       birth:              session[:birth],
       email:              session[:email],
-      password: session[:password],
+      password:           session[:password],
       password_confirmation: session[:password_confirmation],
       first_name:         session[:first_name],
       last_name:          session[:last_name],
@@ -99,24 +101,39 @@ class SignUpsController < ApplicationController
 
     if params[:user][:auth].match(/\A[0-9]+\z/)
       if params[:user][:auth] == "111111"
-        @user.save
-        sign_in User.find(@user.id)
-        session.delete(:nickname)
-        session.delete(:birth)
-        session.delete(:email)
-        session.delete(:password)
-        session.delete(:password_confirmation)
-        session.delete(:first_name)
-        session.delete(:last_name)
-        session.delete(:first_name_kana)
-        session.delete(:last_name_kana)
-        session.delete(:tel)
-        redirect_to registrate_address_path
+        if @user.save
+          if session[:uid]
+            @sns_auth = SnsAuth.new(
+              provider: session[:provider],
+              uid:      session[:uid],
+              user_id:  @user.id,
+            )
+            @sns_auth.save
+            sign_in User.find(@user.id)
+            session.delete(:nickname)
+            session.delete(:birth)
+            session.delete(:email)
+            session.delete(:password)
+            session.delete(:password_confirmation)
+            session.delete(:first_name)
+            session.delete(:last_name)
+            session.delete(:first_name_kana)
+            session.delete(:last_name_kana)
+            session.delete(:tel)
+            session.delete(:uid)
+            session.delete(:provider)
+            redirect_to registrate_address_path
+            return
+          end
+          redirect_to registrate_address_path
+        else
+          render :user_tel_auth, layout: "application_sub"
+        end
       else
         @msg = "認証番号が一致しません"
         render :user_tel_auth, layout: "application_sub"
       end
-    else  
+    else
       @msg = "認証番号を入力してください"
       render :user_tel_auth, layout: "application_sub"
     end
